% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{connect}
\alias{connect}
\title{Create a connection to a PIC-SURE network.}
\usage{
connect(url, token, verbose = FALSE)
}
\arguments{
\item{url}{The url of the PIC-SURE endpoint.}

\item{token}{A security token to access the endpoint.}

\item{verbose}{(optional) set to TRUE to get debugging output}
}
\description{
Create a connection to a PIC-SURE network.
}
\examples{

#   myconn = picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")

}
