% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{list.resources}
\alias{list.resources}
\title{Get a list of resource UUIDs for the passed PIC-SURE connection.}
\usage{
list.resources(connection, verbose = FALSE)
}
\arguments{
\item{connection}{An established PIC-SURE connection object.}

\item{verbose}{(optional) set to TRUE to get debugging output}
}
\description{
Get a list of resource UUIDs for the passed PIC-SURE connection.
}
\examples{

#   myconn = picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")
#   picsure::list.resources(myconn)

}
