% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\docType{package}
\name{picsure}
\alias{picsure}
\alias{picsure-package}
\title{picsure: A Client Library for PIC-SURE Networks.}
\description{
This package is used to connect to a PIC-SURE network and provide
connection services required by all resource-specific libraries/packages.
}
