% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureClient}
\alias{PicSureClient}
\title{R6 class used to establish connections to a PIC-SURE network.}
\format{\code{\link{PicSureClient}} object.}
\usage{
PicSureClient
}
\value{
Object of \code{\link{R6Class}} with methods for connecting to PIC-SURE network.
}
\description{
R6 class used to establish connections to a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-client}
  \item{\code{new()}}{This method is used to create object of this class.}

  \item{\code{version()}}{This method prints the current package version.}
  \item{\code{connect(url, token)}}{This method returns a new \code{PicSureConnection} object configured to connect to a PIC-SURE Network at the given \code{url} and using the provided \code{token} for security.}}
}

\keyword{data}
