% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureConnection}
\alias{PicSureConnection}
\title{R6 class that represents an establish connection to a PIC-SURE network.}
\format{\code{\link{PicSureConnection}} object.}
\usage{
PicSureConnection
}
\value{
Object of \code{\link{R6Class}} with methods for connecting to PIC-SURE network.
}
\description{
R6 class that represents an establish connection to a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-client}
  \item{\code{new(url, token)}}{This method is used to create object of this class.}

  \item{\code{about(resourceId = FALSE)}}{This method... }
  \item{\code{list()}}{This method... }
  \item{\code{getInfo()}}{This method... }
  \item{\code{getResources()}}{This method... }
  \item{\code{INTERNAL_api_obj(url, token)}}{This is an internal method used to retreve a \code{PicSureConnectionAPI} object which handles all communication/authentication with the PIC-SURE Network for a Resource-level package.}}
}

\keyword{data}
