% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureConnectionAPI}
\alias{PicSureConnectionAPI}
\title{R6 class used by other Resource-oriented packages to communicate in a standard
way with a Resource hosted on a PIC-SURE network.}
\format{\code{\link{PicSureConnectionAPI}} object.}
\usage{
PicSureConnectionAPI
}
\value{
Object of \code{\link{R6Class}} with methods for connecting to PIC-SURE network.
}
\description{
R6 class used by other Resource-oriented packages to communicate in a standard
way with a Resource hosted on a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-client}
  \item{\code{new(url, token)}}{This method is used to create object of this class.}

  \item{\code{info(resource_uuid)}}{This method is used by other packages...}
  \item{\code{search(resource_uuid, query)}}{This method is used by other packages...}
  \item{\code{asynchQuery(resource_uuid, query)}}{This method is used by other packages...}
  \item{\code{synchQuery(resource_uuid, query)}}{This method is used by other packages...}
  \item{\code{queryStatus(resource_uuid, query_uuid)}}{This method is used by other packages...}
  \item{\code{queryResult(resource_uuid, query_uuid)}}{This method is used by other packages...}}
}

\keyword{data}
