% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picsure.R
\name{resource.details}
\alias{resource.details}
\title{Get the details of a PIC-SURE resource.}
\usage{
resource.details(connection, resourceUUID, verbose = FALSE)
}
\arguments{
\item{connection}{A PIC-SURE connection object.}

\item{resourceUUID}{The UUID identity of a Resource hosted via the PIC-SURE connection.}

\item{verbose}{(optional) set to TRUE to get debugging output}
}
\value{
A string containing details of the requested Resource.
}
\description{
Get the details of a PIC-SURE resource.
}
\examples{

#   myconn = picsure::connect(url="http://your.pic-sure.server/PIC-SURE/", token="your-security-token")
#   resources = picsure::list.resources(myconn)
#   picsure::resource.details(connection=myconn, resources[0])

}
